// src/app/models/progres.model.ts
export class Progres {
   
    tBucuresti: number;
    tBotosani: number;
    tBrasov: number;
    tCluj: number;
    tConstanta: number;
    tIalomita: number;
    tIasi: number;
    tMaramures: number;
    tMures: number;
    tPrahova: number;
    tSuceava: number;
    pBucuresti: number;
    pBotosani: number;
    pBrasov: number;
    pCluj: number;
    pConstanta: number;
    pIalomita: number;
    pIasi: number;
    pMaramures: number;
    pMures: number;
    pPrahova: number;
    pSuceava: number;
    
    
  
    constructor(input?: any) {
      Object.assign(this, input);
    }
  }
  